requirements.txt
google-cloud-speech>=2.30.0
google-cloud-translate>=3.16.0
PyQt6>=6.7.0
sounddevice>=0.4.7
rapidfuzz>=3.9.0
python-dotenv>=1.0.1
# Optional semantic matching
sentence-transformers>=3.0.0
